Add VariantStoreExporter
